Change Poll package structure
